requirements.txt
# requirements.txt
# Core Framework
flask==2.3.3
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.20
alembic==1.12.0

# WhatsApp Integration
twilio==8.10.0
requests==2.31.0

# NLP & Text Processing
nltk==3.8.1
regex==2023.10.3

# Image Processing & OCR
pillow==10.0.0
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3

# Web Automation
selenium==4.11.0
webdriver-manager==4.0.1

# Date/Time Processing
python-dateutil==2.8.2

# Utilities
python-multipart==0.0.6
pathlib==1.0.1

# Testing
pytest==7.4.2
pytest-mock==3.11.1